+++ tests/UploadControllerTest.cs
@@ -132,6 +132,8 @@ public class UploadControllerTest
         Assert.AreEqual(21106004, lithology.AlterationId);
         Assert.AreEqual("instruction set Dynamic backing up Lock", lithology.Notes);
         Assert.AreEqual("trace back Peso", lithology.OriginalLithology);
+        Assert.AreEqual(30000018, lithology.GradationId);
+        Assert.AreEqual(15101001, lithology.LithologyTopBedrockId);
         var lithoCodeLists = lithology.LayerCodelists;
         Assert.AreEqual(14, lithoCodeLists.Count);
         lithology = stratigraphy.Layers.First(l => l.FromDepth == 11);
@@ -872,7 +874,7 @@ public class UploadControllerTest
         var minBoreholeId = context.Boreholes.Min(b => b.Id);
         var boreholeCsvFile = GetRandomFile("borehoel.csv");
 
-        long targetSizeInBytes = 210 * 1024 * 1024; // 210MB
+        long targetSizeInBytes = 201 * 1024 * 1024; // 201MB
         byte[] content = new byte[targetSizeInBytes];
         var stream = new MemoryStream(content);
 
@@ -882,7 +884,22 @@ public class UploadControllerTest
 
         Assert.IsInstanceOfType(response.Result, typeof(BadRequestObjectResult));
         BadRequestObjectResult badRequestResult = (BadRequestObjectResult)response.Result!;
-        Assert.AreEqual($"One or more attachment exceed maximum file size of 205000000 bytes.", badRequestResult.Value);
+        Assert.AreEqual($"One or more attachment exceed maximum file size of 200 Mb.", badRequestResult.Value);
+    }
+
+    [TestMethod]
+    public async Task UploadWithMaxValidationErrorsExceededShouldReturnError()
+    {
+        var boreholeCsvFile = GetFormFileByExistingFile("maxValidationErrorsExceeded.csv");
+
+        ActionResult<int> response = await controller.UploadFileAsync(workgroupId: 1, boreholeCsvFile);
+
+        Assert.IsInstanceOfType(response.Result, typeof(ObjectResult));
+        ObjectResult result = (ObjectResult)response.Result!;
+        Assert.AreEqual((int)HttpStatusCode.BadRequest, result.StatusCode);
+
+        ValidationProblemDetails problemDetails = (ValidationProblemDetails)result.Value!;
+        Assert.AreEqual(1000, problemDetails.Errors.Count);
     }
 
     [TestMethod]