+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+## v2.0.422 - 2023-06-02
+
 ### Added
 
 - Lithology csv import.
@@ -11,8 +13,9 @@
 ### Changed
 
 - The hydrotest input mask now supports selecting multiple hydrotest kinds.
-- Enabled all file types as borehole attachments.
+- Enable all file types as borehole attachments.
 - Improved appearance of import UI.
+- Increased maximum of displayed validation errors during borehole import to 1000.
 
 ### Fixed
 
@@ -23,6 +26,7 @@
 - Ensure copied stratigraphy is not primary.
 - Ensure borehole attachments can be managed by any user role.
 - Enable upload of borehole attachments with a size of up to 200 Mb. 
+- Show message when borehole import takes longer than 30 seconds.
 
 ## v2.0.400 - 2023-05-24